+++ sdk/keyvault/Azure.Security.KeyVault.Keys/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 4.2.0-beta.3 (Unreleased)
 
+### Added
+
+- Added `KeyType.OctHsm` to support "oct-HSM" key operations.
 
 ## 4.2.0-beta.2 (2020-10-06)